Ensure the knative.dev/hack package doesn't have any deps
