+++ test/go.mod
@@ -5,6 +5,7 @@ go 1.18
 require (
 	github.com/stretchr/testify v1.8.1
 	github.com/thanhpk/randstr v1.0.4
+	golang.org/x/mod v0.12.0
 )
 
 require (
